fix(db): apply parseFlexibleDate everywhere agent-sourced dates land in Prisma writes
Same Invalid-Date class of bug as the company-create date crash, but
this time on director DOB:

  Invalid prisma.director.createMany() invocation:
    dob: new Date("Invalid Date")
  Invalid value for argument `dob`: Provided Date object is invalid.
  Expected Date.

Director DOBs come from the Perplexity / MCA lookup as "DD/MM/YYYY"
or sometimes "Not Available" — neither parses with native new Date().
The previous fix only covered company date fields; the same trap
exists on every repo write that takes an agent-sourced date string.

Promoted parseFlexibleDate from PrismaCompanyRepository into
lib/utils/parse-date.ts and applied it to every at-risk write:

  - Company.create:      incorporation_date, date_of_last_agm,
                         balance_sheet_date  (already covered, now via
                         the shared util)
  - Company.update:      same three fields
  - Director.createMany: dob   (the immediate crash)
  - CompanyDocument.create: registration_date, expiry_date,
                            period_start, period_end   (defensive —
                            agent fills these too)

parseFlexibleDate handles:
  - DD/MM/YYYY and DD-MM-YYYY (Indian default)
  - YYYY/MM/DD and YYYY-MM-DD
  - ISO 8601 via native Date
  - Garbage like "N/A", "Not available", "None", "Unknown", "-"
  - Returns null on anything unparseable — never an Invalid Date

Server-controlled dates (Date.now() in subscription / payment paths)
were left alone — they always produce valid Dates.
